fix: [ANDROAPP-6535] move classes
