+++ app/src/main/java/org/dhis2/usescases/teiDashboard/dashboardfragments/relationships/MapButtonObservable.kt
@@ -1,7 +1,7 @@
 package org.dhis2.usescases.teiDashboard.dashboardfragments.relationships
 
 import androidx.lifecycle.LiveData
-import org.dhis2.tracker.relationships.model.RelationshipTopBarIconState
+import org.dhis2.tracker.relationships.ui.state.RelationshipTopBarIconState
 
 interface MapButtonObservable {
     fun relationshipMap(): LiveData<Boolean>
